Data.ExponentialFamily.Density

Description

Probability Density and related functions.

Documentation

type Probability = Double

class ProbDensity d where

Minimal complete definition

f

Associated Types

type Point d

Methods

f :: d -> Point d -> Probability

f(x|θ), assuming θ is already given in d. Point d is usually Double, but sometimes it may be some other structures.

logF :: d -> Point d -> Double

log (f(x|θ). Handy, nothing more.

expectValD

Arguments

:: (Double -> Probability)
probability density function

-> (Double -> Double)
Gives the value at that

-> Double
Lower

-> Double
Upper

-> Double
dx

-> Double

Specialised, handy version of expectVal.

expectVal

Arguments

:: (ProbDensity d, IntegrateConfig iconf, Point d ~ Domain iconf)

=> iconf
the points and width of the integration

-> d
probability density function

-> (Point d -> Double)
Gives the value at that point

-> Double
expectation

Calculates the expected value of a function under a distribution.

kullbackLeiblerDivergence :: (ProbDensity d0, ProbDensity d1, IntegrateConfig iconf, Point d0 ~ Point d1, Point d0 ~ Domain iconf) => iconf -> d0 -> d1 -> Double

Kullback-Leibler divergence. FIXME : when IntegrateConfig range is too large, it can result in NaN.