Data.ExponentialFamily.Integration

Description

Generic interface for integration.

Documentation

class IntegrateConfig i where

Information about how to calculate the integral. Domain i is usually Double, but we can define integration of some function f : ℂ → ℝ on a line, in which case Domain i might be Complex. We would like to use this to calculate the expected value over discrete random distribution too, for which Domain i shall be an iteration over the support.

Minimal complete definition

integrateValues, integrateWidth

Associated Types

type Domain i

We will calculate integral of f : (Domain i) -> Double.

Methods

integrateValues :: i -> [Domain i]

See the implementation of IntegrateDouble to get grasp of this.

integrateWidth :: i -> Double

integrate

Arguments

:: IntegrateConfig i

=> i
How?

-> (Domain i -> Double)
function to integrate.

-> Double

Integrate a function. non-Double valued functions are currently not supported.

data IntegrateDouble

Integrate from _lower to _upper, using boxes whose width is _dx.

>>> take 5 . show $ integrate (IntegrateDouble 0 (pi/2) 0.001) $ sin
"1.000"

Constructors

IntegrateDouble

Fields

_lower :: Double

_upper :: Double

_dx :: Double

data IntegrateDiscrete a

For Discrete distributions, use all the values with equal weight, 1.

>>> data Foo = Foo | Bar | Baz | Foz deriving (Enum, Bounded)
>>> integrate (IntegrateDiscrete :: IntegrateDiscrete Foo) (fromIntegral . fromEnum)
6.0

Constructors

IntegrateDiscrete