Data.ExponentialFamily.ThetaEta

Description

Anything related to θ-η-coordinates.

Documentation

class ThetaEta a where

Anything that can be expressed with θ- and η- coodinates. We don't use another type for the parameters other than the distribution itself. e.g. We will use Normal μ σ directry and we don't use (μ, σ).

Minimal complete definition

toθ, fromθ, toη, fromη

Associated Types

type TEParam a

TEParam dist is the type for θ- and η- coordinates. For distributions with 2 degrees of freedom, we'll prefer (Double, Double).

Methods

toθ :: a -> Theta a

Get the θ coordinates.

fromθ :: Theta a -> a

Given θ, returns the distribution.

toη :: a -> Eta a

Get the η coordinates.

fromη :: Eta a -> a

Given η, returns the distribution.

modifyθ :: (TEParam a -> TEParam a) -> a -> a

Returns a new distribution with θ modified as instructed.

modifyη :: (TEParam a -> TEParam a) -> a -> a

Returns a new distribution with η modified as instructed.

data Theta a

Theta a and Eta a contains the same thing, usually n Doubles, and we'll use newtypes to make our life a little more type-safe.

Constructors

ThetaEta a => Theta (TEParam a)

θmap :: (ThetaEta a, ThetaEta b) => (TEParam a -> TEParam b) -> Theta a -> Theta b

unθ :: ThetaEta a => Theta a -> TEParam a

data Eta a

Constructors

ThetaEta a => Eta (TEParam a)

ηmap :: (ThetaEta a, ThetaEta b) => (TEParam a -> TEParam b) -> Eta a -> Eta b

unη :: ThetaEta a => Eta a -> TEParam a